refactor(kernel): move AetherSchemaView
